refactor: update mock validators entries_left/entry_limit from u8 to u32

+++ packages/metagame/src/entry_requirement/tests/mocks/accepting_limited_entry_validator_mock.cairo
@@ -71,12 +71,12 @@ pub mod AcceptingLimitedEntryValidatorMock {
             context_id: u64,
             player_address: ContractAddress,
             qualification: Span<felt252>,
-        ) -> Option<u8> {
+        ) -> Option<u32> {
             Option::Some(5)
         }
 
         fn add_config(
-            ref self: ContractState, context_id: u64, entry_limit: u8, config: Span<felt252>,
+            ref self: ContractState, context_id: u64, entry_limit: u32, config: Span<felt252>,
         ) {}
 
         fn add_entry(
